requirements.txt
requests
beautifulsoup4
torch
torchvision
torchaudio
ipykernel
tqdm
numpy
matplotlib 
seaborn 
pandas 
scikit-learn 
scipy
blitz-bayesian-pytorch
